requirements.txt
psycopg2-binary
openai
tiktoken
python-dotenv
pypdf
langchain
numpy